requirements.txt
bs4
pysocks
requests
reportlab
streamlit
python-dotenv
langchain-openai
langchain-ollama
backports.tarfile
langchain-anthropic
langchain_community
langchain_google_genai
selenium
google-api-core